Remove use of `var` and reorganize files
* Moved most auth-related files into com.mystery.project.authentication
* Moved User-related files into com.mystery.project.entities.user
* Removed log statements from AuthenticationService
* Changed nameCheck to emailCheck in AuthenticationService. Previously it was
checking for duplicate display names twice, but now checks for duplicate display
name and duplicate email
* Removed User type variable from register function as it is not necessary to
hold on to
