+++ backend/src/main/java/com/mystery/project/authentication/AuthenticationController.java
@@ -1,34 +1,31 @@
-package com.mystery.project.controllers;
+package com.mystery.project.authentication;
 
+import com.mystery.project.authentication.dtos.LoginRequestDto;
+import com.mystery.project.authentication.dtos.RegisterRequestDto;
 import com.mystery.project.mainconfiguration.Routes;
-import com.mystery.project.services.authentication.AuthenticationService;
-import com.mystery.project.services.authentication.LoginRequestDto;
-import com.mystery.project.services.authentication.RegisterRequestDto;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.web.context.SecurityContextRepository;
 import org.springframework.web.bind.annotation.*;
-
-import java.security.Principal;
 
 @RestController
 @RequestMapping(Routes.AUTHENTICATION)
 @RequiredArgsConstructor
 public class AuthenticationController {
   private final AuthenticationService authenticationService;
-  private final AuthenticationManager authenticationManager;
   private final SecurityContextRepository securityContextRepository;
 
   @PostMapping("/login")
   public void login(@RequestBody LoginRequestDto loginRequestDto, HttpServletRequest request, HttpServletResponse response) {
-    var authenticationResult = authenticationService.loginUser(loginRequestDto);
+    Authentication authenticationResult = authenticationService.loginUser(loginRequestDto);
+    SecurityContext context = SecurityContextHolder.createEmptyContext();
 
-    var context = SecurityContextHolder.createEmptyContext();
     context.setAuthentication(authenticationResult);
     SecurityContextHolder.setContext(context);
     securityContextRepository.saveContext(context, request, response);
@@ -36,8 +33,9 @@ public class AuthenticationController {
 
   @PostMapping("/register")
   public void register(@RequestBody RegisterRequestDto registerRequestDto, HttpServletRequest request, HttpServletResponse response) {
-    var authenticationResult = authenticationService.registerUser(registerRequestDto);
-    var context = SecurityContextHolder.createEmptyContext();
+    Authentication authenticationResult = authenticationService.registerUser(registerRequestDto);
+    SecurityContext context = SecurityContextHolder.createEmptyContext();
+
     context.setAuthentication(authenticationResult);
     SecurityContextHolder.setContext(context);
     securityContextRepository.saveContext(context, request, response);
