+++ backend/src/main/java/com/mystery/project/authentication/dtos/RegisterRequestDto.java
@@ -1,3 +1,3 @@
-package com.mystery.project.services.authentication;
+package com.mystery.project.authentication.dtos;
 
 public record RegisterRequestDto(String email, String password, String displayName) {}
