+++ backend/src/main/java/com/mystery/project/authentication/exceptions/DuplicateEmailException.java
@@ -1,4 +1,4 @@
-package com.mystery.project.exception;
+package com.mystery.project.authentication.exceptions;
 
 public class DuplicateEmailException extends RuntimeException {
   public DuplicateEmailException(String message) {
