+++ backend/src/main/java/com/mystery/project/authentication/exceptions/EmailNotFoundException.java
@@ -1,4 +1,4 @@
-package com.mystery.project.exception;
+package com.mystery.project.authentication.exceptions;
 
 public class EmailNotFoundException extends RuntimeException {
   public EmailNotFoundException(String message) {
