+++ backend/src/main/java/com/mystery/project/securityconfig/SecurityConfig.java
@@ -1,25 +1,19 @@
 package com.mystery.project.securityconfig;
 
+import com.mystery.project.authentication.CustomUserDetailsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.ProviderManager;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
 import org.springframework.security.config.annotation.web.configurers.FormLoginConfigurer;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.context.DelegatingSecurityContextRepository;
-import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
-import org.springframework.security.web.context.RequestAttributeSecurityContextRepository;
 import org.springframework.security.web.context.SecurityContextRepository;
 
 @Configuration
@@ -34,11 +28,9 @@ public class SecurityConfig {
     return http.securityContext(
             securityContext -> securityContext.securityContextRepository(securityContextRepository))
         .authorizeHttpRequests(
-            authorize -> {
-              authorize
-                  .requestMatchers("/api/v1/auth/login", "/api/v1/auth/register").permitAll()
-                  .anyRequest().authenticated();
-            })
+            authorize -> authorize
+                .requestMatchers("/api/v1/auth/login", "/api/v1/auth/register").permitAll()
+                .anyRequest().authenticated())
             .userDetailsService(customUserDetailsService)
         .formLogin(FormLoginConfigurer::disable)
         .csrf(CsrfConfigurer::disable)
