--- backend/src/main/java/com/mystery/project/services/authentication/UserDto.java
@@ -1,5 +1,0 @@
-package com.mystery.project.services.authentication;
-
-import java.util.UUID;
-
-public record UserDto(UUID id, String displayName, String email) {}
